+++ example/lib/player_widget.dart
@@ -18,7 +18,7 @@ class PlayerWidget extends StatefulWidget {
 
   @override
   State<StatefulWidget> createState() {
-    return new _PlayerWidgetState(url, isLocal, mode);
+    return _PlayerWidgetState(url, isLocal, mode);
   }
 }
 
@@ -65,61 +65,61 @@ class _PlayerWidgetState extends State<PlayerWidget> {
 
   @override
   Widget build(BuildContext context) {
-    return new Column(
+    return Column(
       mainAxisSize: MainAxisSize.min,
       children: <Widget>[
-        new Row(
+        Row(
           mainAxisSize: MainAxisSize.min,
           children: [
-            new IconButton(
+            IconButton(
                 onPressed: _isPlaying ? null : () => _play(),
                 iconSize: 64.0,
-                icon: new Icon(Icons.play_arrow),
+                icon: Icon(Icons.play_arrow),
                 color: Colors.cyan),
-            new IconButton(
+            IconButton(
                 onPressed: _isPlaying ? () => _pause() : null,
                 iconSize: 64.0,
-                icon: new Icon(Icons.pause),
+                icon: Icon(Icons.pause),
                 color: Colors.cyan),
-            new IconButton(
+            IconButton(
                 onPressed: _isPlaying || _isPaused ? () => _stop() : null,
                 iconSize: 64.0,
-                icon: new Icon(Icons.stop),
+                icon: Icon(Icons.stop),
                 color: Colors.cyan),
           ],
         ),
-        new Row(
+        Column(
           mainAxisSize: MainAxisSize.min,
           children: [
-            new Padding(
-              padding: new EdgeInsets.all(12.0),
-              child: new Stack(
+            Padding(
+              padding: EdgeInsets.all(12.0),
+              child: Stack(
                 children: [
-                  new CircularProgressIndicator(
-                    value: 1.0,
-                    valueColor: new AlwaysStoppedAnimation(Colors.grey[300]),
-                  ),
-                  new CircularProgressIndicator(
+                  Slider(
+                    onChanged: (v) {
+                      final Position = v * _duration.inMilliseconds;
+                      _audioPlayer
+                          .seek(Duration(milliseconds: Position.round()));
+                    },
                     value: (_position != null &&
                             _duration != null &&
                             _position.inMilliseconds > 0 &&
                             _position.inMilliseconds < _duration.inMilliseconds)
                         ? _position.inMilliseconds / _duration.inMilliseconds
                         : 0.0,
-                    valueColor: new AlwaysStoppedAnimation(Colors.cyan),
                   ),
                 ],
               ),
             ),
-            new Text(
+            Text(
               _position != null
                   ? '${_positionText ?? ''} / ${_durationText ?? ''}'
                   : _duration != null ? _durationText : '',
-              style: new TextStyle(fontSize: 24.0),
+              style: TextStyle(fontSize: 24.0),
             ),
           ],
         ),
-        new Text("State: $_audioPlayerState")
+        Text("State: $_audioPlayerState")
       ],
     );
   }
@@ -127,10 +127,23 @@ class _PlayerWidgetState extends State<PlayerWidget> {
   void _initAudioPlayer() {
     _audioPlayer = AudioPlayer(mode: mode);
 
-    _durationSubscription =
-        _audioPlayer.onDurationChanged.listen((duration) => setState(() {
-              _duration = duration;
-            }));
+    _durationSubscription = _audioPlayer.onDurationChanged.listen((duration) {
+      setState(() => _duration = duration);
+
+      // TODO implemented for iOS, waiting for android impl
+      if (Theme.of(context).platform == TargetPlatform.iOS) {
+        // set atleast title to see the notification bar on ios.
+        _audioPlayer.setNotification(
+            title: 'App Name',
+            artist: 'Artist or blank',
+            albumTitle: 'Name or blank',
+            imageUrl: 'url or blank',
+            forwardSkipInterval: const Duration(seconds: 30), // default is 30s
+            backwardSkipInterval: const Duration(seconds: 30), // default is 30s
+            duration: duration,
+            elapsedTime: Duration(seconds: 0));
+      }
+    });
 
     _positionSubscription =
         _audioPlayer.onAudioPositionChanged.listen((p) => setState(() {
@@ -160,6 +173,11 @@ class _PlayerWidgetState extends State<PlayerWidget> {
         _audioPlayerState = state;
       });
     });
+
+    _audioPlayer.onNotificationPlayerStateChanged.listen((state) {
+      if (!mounted) return;
+      setState(() => _audioPlayerState = state);
+    });
   }
 
   Future<int> _play() async {
@@ -172,6 +190,15 @@ class _PlayerWidgetState extends State<PlayerWidget> {
     final result =
         await _audioPlayer.play(url, isLocal: isLocal, position: playPosition);
     if (result == 1) setState(() => _playerState = PlayerState.playing);
+
+    // TODO implemented for iOS, waiting for android impl
+    if (Theme.of(context).platform == TargetPlatform.iOS) {
+      // default playback rate is 1.0
+      // this should be called after _audioPlayer.play() or _audioPlayer.resume()
+      // this can also be called everytime the user wants to change playback rate in the UI
+      _audioPlayer.setPlaybackRate(playbackRate: 1.0);
+    }
+
     return result;
   }