+++ lib/audioplayers.dart
@@ -4,6 +4,10 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:uuid/uuid.dart';
 
+// Required for PluginUtilities.
+import 'dart:ui';
+import 'package:flutter/material.dart';
+
 typedef StreamController CreateStreamController();
 typedef void TimeChangeHandler(Duration duration);
 typedef void ErrorHandler(String message);
@@ -59,6 +63,50 @@ enum PlayerMode {
   LOW_LATENCY
 }
 
+// When we start the background service isolate, we only ever enter it once.
+// To communicate between the native plugin and this entrypoint, we'll use
+// MethodChannels to open a persistent communication channel to trigger
+// callbacks.
+void _backgroundCallbackDispatcher() {
+  const MethodChannel _channel =
+      MethodChannel('xyz.luan/audioplayers_callback');
+
+  // Setup Flutter state needed for MethodChannels.
+  WidgetsFlutterBinding.ensureInitialized();
+
+  // Reference to the onAudioChangeBackgroundEvent callback.
+  Function onAudioChangeBackgroundEvent;
+
+  // This is where the magic happens and we handle background events from the
+  // native portion of the plugin. Here we message the audio notification data 
+  // which we then pass to the provided callback.
+  _channel.setMethodCallHandler((MethodCall call) async {
+    Function _performCallbackLookup() {
+      final CallbackHandle handle = CallbackHandle.fromRawHandle(
+          call.arguments['updateHandleMonitorKey']);
+
+      // PluginUtilities.getCallbackFromHandle performs a lookup based on the
+      // handle we retrieved earlier.
+      final Function closure = PluginUtilities.getCallbackFromHandle(handle);
+
+      if (closure == null) {
+        print('Fatal Error: Callback lookup failed!');
+        // exit(-1);
+      }
+      return closure;
+    }
+
+    final Map<dynamic, dynamic> callArgs = call.arguments as Map;
+    if (call.method == 'audio.onNotificationBackgroundPlayerStateChanged') {
+      onAudioChangeBackgroundEvent ??= _performCallbackLookup();
+      final bool isPlaying = callArgs['value'];
+      onAudioChangeBackgroundEvent(isPlaying);
+    } else {
+      assert(false, "No handler defined for method type: '${call.method}'");
+    }
+  });
+}
+
 /// This represents a single AudioPlayer, which can play one audio at a time.
 /// To play several audios at the same time, you must create several instances
 /// of this class.
@@ -75,6 +123,9 @@ class AudioPlayer {
   final StreamController<AudioPlayerState> _playerStateController =
       StreamController<AudioPlayerState>.broadcast();
 
+  final StreamController<AudioPlayerState> _notificationPlayerStateController =
+      StreamController<AudioPlayerState>.broadcast();
+
   final StreamController<Duration> _positionController =
       StreamController<Duration>.broadcast();
 
@@ -83,6 +134,9 @@ class AudioPlayer {
 
   final StreamController<void> _completionController =
       StreamController<void>.broadcast();
+
+  final StreamController<void> _seekCompleteController =
+  StreamController<void>.broadcast();
 
   final StreamController<String> _errorController =
       StreamController<String>.broadcast();
@@ -107,9 +161,18 @@ class AudioPlayer {
     _audioPlayerState = state;
   }
 
+  set notificationState(AudioPlayerState state) {
+    _notificationPlayerStateController.add(state);
+    _audioPlayerState = state;
+  }
+
   /// Stream of changes on player state.
   Stream<AudioPlayerState> get onPlayerStateChanged =>
       _playerStateController.stream;
+
+  /// Stream of changes on player state coming from notification area in iOS.
+  Stream<AudioPlayerState> get onNotificationPlayerStateChanged =>
+      _notificationPlayerStateController.stream;
 
   /// Stream of changes on audio position.
   ///
@@ -132,6 +195,11 @@ class AudioPlayer {
   ///
   /// [ReleaseMode.LOOP] also sends events to this stream.
   Stream<void> get onPlayerCompletion => _completionController.stream;
+
+  /// Stream of seek completions.
+  ///
+  /// An event is going to be sent as soon as the audio seek is finished.
+  Stream<void> get onSeekComplete => _seekCompleteController.stream;
 
   /// Stream of player errors.
   ///
@@ -173,6 +241,14 @@ class AudioPlayer {
   @deprecated
   VoidCallback completionHandler;
 
+  /// Handler of seek completion.
+  ///
+  /// An event is going to be sent as soon as the audio seek is finished.
+  ///
+  /// This is deprecated. Use [onSeekComplete] instead.
+  @deprecated
+  VoidCallback seekCompleteHandler;
+
   /// Handler of player errors.
   ///
   /// Events are sent when an unexpected error is thrown in the native code.
@@ -191,10 +267,20 @@ class AudioPlayer {
   PlayerMode mode;
 
   /// Creates a new instance and assigns an unique id to it.
-  AudioPlayer({this.mode = PlayerMode.MEDIA_PLAYER}) {
+  AudioPlayer({this.mode = PlayerMode.MEDIA_PLAYER, this.playerId}) {
     this.mode ??= PlayerMode.MEDIA_PLAYER;
-    playerId = _uuid.v4();
+    this.playerId ??= _uuid.v4();
     players[playerId] = this;
+
+    // Start the headless location service. The parameter here is a handle to
+    // a callback managed by the Flutter engine, which allows for us to pass
+    // references to our callbacks between isolates.
+    final CallbackHandle handle =
+        PluginUtilities.getCallbackHandle(_backgroundCallbackDispatcher);
+    assert(handle != null, 'Unable to lookup callback.');
+    _invokeMethod('startHeadlessService', {
+      'handleKey': <dynamic>[handle.toRawHandle()]
+    });
   }
 
   Future<int> _invokeMethod(
@@ -210,6 +296,24 @@ class AudioPlayer {
     return _channel
         .invokeMethod(method, withPlayerId)
         .then((result) => (result as int));
+  }
+
+  /// Start getting significant location updates through `callback`.
+  ///
+  /// `callback` is invoked on a background isolate and will not have direct
+  /// access to the state held by the main isolate (or any other isolate).
+  Future<bool> monitorNotificationStateChanges(
+      void Function(bool value) callback) async {
+    if (callback == null) {
+      throw ArgumentError.notNull('callback');
+    }
+    final CallbackHandle handle = PluginUtilities.getCallbackHandle(callback);
+
+    await _invokeMethod('monitorNotificationStateChanges', {
+      'handleMonitorKey': <dynamic>[handle.toRawHandle()]
+    });
+
+    return true;
   }
 
   /// Plays an audio.
@@ -304,6 +408,7 @@ class AudioPlayer {
 
   /// Moves the cursor to the desired position.
   Future<int> seek(Duration position) {
+    _positionController.add(position);
     return _invokeMethod('seek', {'position': position.inMilliseconds});
   }
 
@@ -325,14 +430,48 @@ class AudioPlayer {
     );
   }
 
+  /// Sets the playback rate - call this after first calling play() or resume(). Works only on iOS for now
+  ///
+  /// iOS has limits between 0.5 and 2x
+  /// not sure if that's changed recently.
+  Future<int> setPlaybackRate({double playbackRate = 1.0}) {
+    return _invokeMethod('setPlaybackRate', {'playbackRate': playbackRate});
+  }
+
+  /// Sets the notification bar for lock screen and notification area in iOS for now.
+  ///
+  /// Specify atleast title
+  Future<dynamic> setNotification(
+      {String title,
+      String albumTitle = '',
+      String artist = '',
+      String imageUrl = '',
+      Duration forwardSkipInterval = const Duration(seconds: 30),
+      Duration backwardSkipInterval = const Duration(seconds: 30),
+      Duration duration,
+      Duration elapsedTime}) {
+    return _invokeMethod('setNotification', {
+      'title': title,
+      'albumTitle': albumTitle,
+      'artist': artist,
+      'imageUrl': imageUrl,
+      'forwardSkipInterval': forwardSkipInterval?.inSeconds ?? 30,
+      'backwardSkipInterval': backwardSkipInterval?.inSeconds ?? 30,
+      'duration': duration?.inSeconds ?? 0,
+      'elapsedTime': elapsedTime?.inSeconds ?? 0
+    });
+  }
+
   /// Sets the URL.
   ///
   /// Unlike [play], the playback will not resume.
   ///
   /// The resources will start being fetched or buffered as soon as you call
   /// this method.
-  Future<int> setUrl(String url, {bool isLocal: false}) {
-    return _invokeMethod('setUrl', {'url': url, 'isLocal': isLocal});
+  Future<int> setUrl(String url,
+      {bool isLocal: false, bool respectSilence = false}) {
+    return _invokeMethod('setUrl',
+        {'url': url, 'isLocal': isLocal, 'respectSilence': respectSilence});
   }
 
   /// Get audio duration after setting url.
@@ -366,6 +505,11 @@ class AudioPlayer {
     final value = callArgs['value'];
 
     switch (call.method) {
+      case 'audio.onNotificationPlayerStateChanged':
+        final bool isPlaying = value;
+        player.notificationState =
+            isPlaying ? AudioPlayerState.PLAYING : AudioPlayerState.PAUSED;
+        break;
       case 'audio.onDuration':
         Duration newDuration = Duration(milliseconds: value);
         player._durationController.add(newDuration);
@@ -383,6 +527,10 @@ class AudioPlayer {
         player._completionController.add(null);
         // ignore: deprecated_member_use_from_same_package
         player.completionHandler?.call();
+        break;
+      case 'audio.onSeekComplete':
+        player._seekCompleteController.add(null);
+        player.seekCompleteHandler?.call();
         break;
       case 'audio.onError':
         player.state = AudioPlayerState.STOPPED;
@@ -410,10 +558,13 @@ class AudioPlayer {
 
     if (!_playerStateController.isClosed)
       futures.add(_playerStateController.close());
+    if (!_notificationPlayerStateController.isClosed)
+      futures.add(_notificationPlayerStateController.close());
     if (!_positionController.isClosed) futures.add(_positionController.close());
     if (!_durationController.isClosed) futures.add(_durationController.close());
     if (!_completionController.isClosed)
       futures.add(_completionController.close());
+    if (!_seekCompleteController.isClosed) futures.add(_seekCompleteController.close());
     if (!_errorController.isClosed) futures.add(_errorController.close());
 
     await Future.wait(futures);